[fix] Allow empty body as null in confirm (#113)
## Before this PR

One couldn't send an empty-body `/confirm/...` request to an endpoint that has a nullable body (e.g. `receiveOptionalBearerTokenAliasExample`, or any other optional / collection endpoint).
This would be met with an `"Default:UnsupportedContentType"` error.

## After this PR

Empty body requests to `/confirm/...`  are treated as if the JSON value `null` was sent.
If a content type was set, it's expected that it was JSON, otherwise we fail.
This allows the conjure serde logic to deserialize the null into an empty collection / absent optional, if the conjure type permits, or otherwise to fail as expected.